fix(storage): 修复 clippy unnecessary_cast 警告 + 更新 proptest 回归基准
- cas.rs:66 Linux 版本下 f_bavail/f_frsize 已是 u64，移除冗余 as u64 转换
- 更新 proptest-regressions/storage/metadata_store.txt 至最新测试基准

+++ log-analyzer/src-tauri/src/storage/cas.rs
@@ -62,8 +62,8 @@ async fn get_available_space(path: &Path) -> u64 {
 
     if result == 0 {
         // f_bavail: 非超级用户可用的块数
-        // f_frsize: 每个块的字节数
-        (stat.f_bavail as u64) * (stat.f_frsize as u64)
+        // f_frsize: 每个块的字节数（Linux 下已是 u64，无需转换）
+        stat.f_bavail * stat.f_frsize
     } else {
         0
     }
